% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUE_BIOMASS.r
\docType{data}
\name{LUE_BIOMASS}
\alias{LUE_BIOMASS}
\title{Light Use efficiency model to estimate biomass}
\format{A Biomass raster}
\usage{
LUE_BIOMASS(fpar_raster,par,tmin,tmin_min,tmin_max,LUE_optimal)
}
\arguments{
\item{fpar_raster}{fraction of photosynthetically active radiation (fpar) per day raster with .tif format}

\item{par}{clear sky surface photosynthetically active radiation (par) per day raster with .nc file format.}

\item{tmin}{Minimum temperature at 2 metres since previous post-processing per day raster with .nc file format.}

\item{tmin_min}{minimum value of tmin used for the threshold}

\item{tmin_max}{maximum value of tmin used for the threshold}

\item{LUE_optimal}{optical lue value with respect to crop type for example wheat crop LUE_optimal is 3.0 (Djumaniyazova et al., 2010)}
}
\value{
Biomass raster
}
\description{
Contains LUE_BIOMASS function to estimate aboveground biomass firstly by calculating the apar and then the minimum temperature and actual values of light use efficiency are further used to calculate the biomass of a crop.
}
\examples{
\dontrun{
## load the data
fpar<-load("~/R/Maninder_eagles/lue/data/fpar.rda")
par1<-load("~/R/Maninder_eagles/lue/data/par1.rda")
tmin<-load("~/R/Maninder_eagles/lue/data/tmin.rda")
LUE_BIOMASS(fpar,par1,tmin,-2,12,3)
}
}
\references{
Djumaniyazova Y, Sommer R, Ibragimov N, Ruzimov J, Lamers J & Vlek P (2010) Simulating water use and N response of winter wheat in the irrigated floodplains of Northwest Uzbekistan. Field Crops Research 116, 239-251.

Shi Z, Ruecker G R,Mueller M, Conrad C, Ibragimov N, Lamers J P A, Martius C, Strunz G, Dech S & Vlek P L G (2007) Modeling of Cotton Yields in the Amu Darya River Floodplains of Uzbekistan Integrating Multitemporal Remote Sensing and Minimum Field Data. Agronomy Journal 99, 1317-1326
}
\keyword{datasets}
